requirements.txt
antlr4-python3-runtime==4.7.2
astroid==2.2.5
autopep8==1.4.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
